Renamed commands to commandbar
