+++ src/frontend/js/views/RightCommandBar.jsx
